+++ projects/core/src/entity.ts
@@ -192,7 +192,19 @@ export interface EntityOptions<entityType = any> {
    * dbName:'public."myProducts"'
    */
   dbName?: string
-  /** For entities that are based on SQL expressions instead of a physical table or view*/
+  /** For entities that are based on SQL expressions instead of a physical table or view
+   * @example
+   * .@Entity('people',{
+   * sqlExpression:`select id,name from employees
+   *      union all select id,name from contractors`,
+   * })
+   * export class Person{
+   * .@Fields.string()
+   * id=''
+   * .@Fields.string()
+   * name=''
+   * }
+   */
   sqlExpression?:
     | string
     | ((entity: EntityMetadata<entityType>) => string | Promise<string>)